Fix #5667 to add native support for camel-quarkus-xj
