+++ docs/modules/ROOT/examples/components/xj.yml
@@ -2,11 +2,11 @@
 # This file was generated by camel-quarkus-maven-plugin:update-extension-doc-page
 cqArtifactId: camel-quarkus-xj
 cqArtifactIdBase: xj
-cqNativeSupported: false
-cqStatus: Preview
+cqNativeSupported: true
+cqStatus: Stable
 cqDeprecated: false
 cqJvmSince: 1.1.0
-cqNativeSince: n/a
+cqNativeSince: 3.7.0
 cqCamelPartName: xj
 cqCamelPartTitle: XJ
 cqCamelPartDescription: Transform JSON and XML message using a XSLT.
