+++ extensions/pom.xml
@@ -246,6 +246,7 @@
         <module>vertx-websocket</module>
         <module>weather</module>
         <module>xchange</module>
+        <module>xj</module>
         <module>xmlsecurity</module>
         <module>xpath</module>
         <module>xslt</module>
