+++ extensions/xj/runtime/src/main/resources/META-INF/quarkus-extension.yaml
@@ -1,0 +1,33 @@
+#
+# Licensed to the Apache Software Foundation (ASF) under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.
+# The ASF licenses this file to You under the Apache License, Version 2.0
+# (the "License"); you may not use this file except in compliance with
+# the License.  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+# This is a generated file. Do not edit directly!
+# To re-generate, run the following command from the top level directory:
+#
+#   mvn -N cq:update-quarkus-metadata
+#
+---
+name: "Camel XJ"
+description: "Transform JSON and XML message using a XSLT"
+metadata:
+  icon-url: "https://camel.apache.org/_/img/logo-d-f21b25ba38.svg"
+  unlisted: true
+  guide: "https://camel.apache.org/camel-quarkus/latest/reference/extensions/xj.html"
+  categories:
+  - "integration"
+  status:
+  - "stable"
