+++ integration-tests-jvm/pom.xml
@@ -97,7 +97,6 @@
         <module>web3j</module>
         <module>wordpress</module>
         <module>workday</module>
-        <module>xj</module>
         <module>xml-grouped</module>
         <module>xmpp</module>
         <module>zookeeper</module>
