+++ integration-tests/xj/src/main/java/org/apache/camel/quarkus/component/xj/it/XjResource.java
