+++ integration-tests/xj/src/test/java/org/apache/camel/quarkus/component/xj/it/XjTest.java
